requirements.txt
pyyaml>=3.1.0
docopt
schema
reportlab
pypdf2
networkx
psutil
curio==0.9
matplotlib
tqdm
oauth2
dateparser
pytz
textblob
msal
atexit
pyyaml-include